avance2

+++ prometeo/src/main/java/com/prometeo/application/entity/estadistica/Discreta.java
@@ -4,8 +4,8 @@ package com.prometeo.application.entity.estadistica;
  *
  * @author L
  */
-public non-sealed class Discreta extends Variable<Integer>{
-    public Discreta(String nombre, Integer valor) {
+public non-sealed class Discreta extends Variable<Double>{
+    public Discreta(String nombre, Double valor) {
         super(nombre, SubTypeVariable.DISCRETA, valor);
     }
 }
